requirements.txt
pyttsx3==2.90
SpeechRecognition>=3.8.1
keyboard==0.13.5
python-decouple==3.8
wikipedia>=1.4.0
pywhatkit>=5.4
requests>=2.32.3